refactor():컴포넌트 파일구조 최적화
1.컴포넌트를 ui기능단위로 폴더를 만들어 폴더구조 개선

2.에어비앤비 컨벤션에 따라 각 폴더 내 대표적인 컴포넌트 파일 명 index.jsx로 변경

3.에어비앤비 컨벤션에 따라 컴포넌트 명, 파일 네이밍 준수
