+++ src/components/Main/RestaurantCategoryFilter/RestaurantCategoryFilter.module.css
@@ -16,7 +16,3 @@
 
   font-size: 16px;
 }
-
-.restaurantFilter {
-  padding: 8px;
-}
