+++ src/components/Main/RestaurantCategoryFilter/index.jsx
@@ -1,5 +1,5 @@
 import styles from './RestaurantCategoryFilter.module.css';
-import Filter from '../UI/filter';
+import Filter from '../../UI/Filter';
 
 function RestaurantCategoryFilter({ restaurants, setfilteredRestaurants }) {
   const CATEGORIES = ['전체', '한식', '중식', '일식', '양식', '아시안', '기타'];
