+++ src/components/UI/Filter/Filter.module.css
@@ -1,0 +1,3 @@
+.restaurantFilter {
+  padding: 8px;
+}
