+++ src/components/UI/Filter/index.jsx
@@ -1,4 +1,4 @@
-import styles from './filter.module.css';
+import styles from './Filter.module.css';
 
 function Filter({ label, options, onChange }) {
   return (
